create Quit(Image/button) from account with confirmation

+++ SecurePass/ViewModels/MainWindowVM.cs
@@ -54,8 +54,8 @@ namespace SecurePass.ViewModels
             isDescendingSort;
         private readonly UnitOfWork repository;
         private UserVM? currentUser;
-        private string? findString, oldPassword, newPassword;
-        private CategoryVM? selectedCategory, categoryInObjectView;
+        private string? findString,userPassword,userLogin,oldPassword,newPassword;
+        private CategoryVM? selectedCategory,categoryInObjectView;
         private List<SecureObjectVM> secureObjects = new();
         private SecureObjectVM? selectedSecureObject, secureObjectEdit;
         private BaseEntityVM? createdObject;
@@ -102,18 +102,27 @@ namespace SecurePass.ViewModels
 
         private void clearData()
         {
-            createdObject = null;
-            isMainWindowEnabled = false;
-            isFirstStart = false;
-            secureObjects.Clear();
-            UserCategories.Clear();
-            UserLogin = string.Empty;
-            UserPassword = string.Empty;
-            CurrentUser = null;
-            SelectedCategory = null;
-            SelectedSecureObject = null;
-            findString = string.Empty;
-            RegistryUtility.DeleteInfoFromRegistry();
+            MessageBoxResult result = MessageBox.Show("Are you sure you want to log out?", "Logout confirmation", MessageBoxButton.YesNo, MessageBoxImage.Question);
+
+            if (result == MessageBoxResult.Yes)
+            {
+                createdObject = null;
+                secureObjects.Clear();
+                UserCategories.Clear();
+                UserLogin = null;
+                UserPassword = null;
+                CurrentUser = null;
+                SelectedCategory = null;
+                SelectedSecureObject = null;
+                findString = string.Empty;
+                RegistryUtility.DeleteInfoFromRegistry();
+                IsFirstStart = true;
+                IsMainWindowEnabled = false;
+            }
+            else
+            {
+
+            }
         }
 
         private async Task setObjectToDataBase(BaseEntityVM? vm)
@@ -902,10 +911,27 @@ namespace SecurePass.ViewModels
         }
 
         // User login value 
-        public string UserLogin { get; set; } = string.Empty;
+        public string UserLogin
+        {
+            get => userLogin;
+            set
+            {
+                userLogin = value;
+                OnPropertyChanged();
+            }
+        }
+
 
         // User password value 
-        public string UserPassword { get; set; } = string.Empty;
+        public string UserPassword 
+        { 
+            get => userPassword; 
+            set 
+            {
+                userPassword = value;
+                OnPropertyChanged();
+            }
+        }
 
         public IEnumerable<ImageVM> DefaultCategoryImages => ImageLoader.DefaultCategoryImages.Select(x => new ImageVM(x));
 
@@ -920,6 +946,8 @@ namespace SecurePass.ViewModels
         public RelayCommand AddNewObject => new( (o) => IsAddObjectWindowEnabled = !IsAddObjectWindowEnabled ,(o)=> UserCategories.Count != 0);
         public RelayCommand ChangeImage => new((o) => changeImage(o), (o) => (o is not SecureObjectVM) || (o as SecureObjectVM).IsEditable);
         public RelayCommand SetCategoryImage => new ((o) => setCategoryImage(o));
+        public RelayCommand QuitFromAccount => new((object o) => clearData());
         public RelayCommand Sort => new((o) => IsDescendingSort = !IsDescendingSort);
+
     }
 }
